More reorganization and renaming.
